// src/lib/api/models/ApiKeyPublic.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ApiKeyPublic
 */
export interface ApiKeyPublic {
    /**
     * 
     * @type {string}
     * @memberof ApiKeyPublic
     */
    name: string;
    /**
     * 
     * @type {boolean}
     * @memberof ApiKeyPublic
     */
    isActive?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ApiKeyPublic
     */
    expiresAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ApiKeyPublic
     */
    permissions?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiKeyPublic
     */
    id: string;
    /**
     * 
     * @type {Date}
     * @memberof ApiKeyPublic
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ApiKeyPublic
     */
    userId: string;
    /**
     * 
     * @type {string}
     * @memberof ApiKeyPublic
     */
    prefix: string;
    /**
     * 
     * @type {string}
     * @memberof ApiKeyPublic
     */
    plainKey?: string | null;
}

/**
 * Check if a given object implements the ApiKeyPublic interface.
 */
export function instanceOfApiKeyPublic(value: object): value is ApiKeyPublic {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('userId' in value) || value['userId'] === undefined) return false;
    if (!('prefix' in value) || value['prefix'] === undefined) return false;
    return true;
}

export function ApiKeyPublicFromJSON(json: any): ApiKeyPublic {
    return ApiKeyPublicFromJSONTyped(json, false);
}

export function ApiKeyPublicFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiKeyPublic {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'isActive': json['is_active'] == null ? undefined : json['is_active'],
        'expiresAt': json['expires_at'] == null ? undefined : (new Date(json['expires_at'])),
        'permissions': json['permissions'] == null ? undefined : json['permissions'],
        'id': json['id'],
        'createdAt': (new Date(json['created_at'])),
        'userId': json['user_id'],
        'prefix': json['prefix'],
        'plainKey': json['plain_key'] == null ? undefined : json['plain_key'],
    };
}

export function ApiKeyPublicToJSON(json: any): ApiKeyPublic {
    return ApiKeyPublicToJSONTyped(json, false);
}

export function ApiKeyPublicToJSONTyped(value?: ApiKeyPublic | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'is_active': value['isActive'],
        'expires_at': value['expiresAt'] == null ? value['expiresAt'] : value['expiresAt'].toISOString(),
        'permissions': value['permissions'],
        'id': value['id'],
        'created_at': value['createdAt'].toISOString(),
        'user_id': value['userId'],
        'prefix': value['prefix'],
        'plain_key': value['plainKey'],
    };
}