// src/lib/api/models/ApiKeysPublic.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ApiKeyPublic } from './ApiKeyPublic';
import {
    ApiKeyPublicFromJSON,
    ApiKeyPublicFromJSONTyped,
    ApiKeyPublicToJSON,
    ApiKeyPublicToJSONTyped,
} from './ApiKeyPublic';

/**
 * 
 * @export
 * @interface ApiKeysPublic
 */
export interface ApiKeysPublic {
    /**
     * 
     * @type {Array<ApiKeyPublic>}
     * @memberof ApiKeysPublic
     */
    data: Array<ApiKeyPublic>;
    /**
     * 
     * @type {number}
     * @memberof ApiKeysPublic
     */
    count: number;
}

/**
 * Check if a given object implements the ApiKeysPublic interface.
 */
export function instanceOfApiKeysPublic(value: object): value is ApiKeysPublic {
    if (!('data' in value) || value['data'] === undefined) return false;
    if (!('count' in value) || value['count'] === undefined) return false;
    return true;
}

export function ApiKeysPublicFromJSON(json: any): ApiKeysPublic {
    return ApiKeysPublicFromJSONTyped(json, false);
}

export function ApiKeysPublicFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiKeysPublic {
    if (json == null) {
        return json;
    }
    return {
        
        'data': ((json['data'] as Array<any>).map(ApiKeyPublicFromJSON)),
        'count': json['count'],
    };
}

export function ApiKeysPublicToJSON(json: any): ApiKeysPublic {
    return ApiKeysPublicToJSONTyped(json, false);
}

export function ApiKeysPublicToJSONTyped(value?: ApiKeysPublic | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': ((value['data'] as Array<any>).map(ApiKeyPublicToJSON)),
        'count': value['count'],
    };
}