// src/lib/api/models/ApiKeyCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ApiKeyCreate
 */
export interface ApiKeyCreate {
    /**
     * 
     * @type {string}
     * @memberof ApiKeyCreate
     */
    name: string;
    /**
     * 
     * @type {boolean}
     * @memberof ApiKeyCreate
     */
    isActive?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ApiKeyCreate
     */
    expiresAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ApiKeyCreate
     */
    permissions?: string;
}

/**
 * Check if a given object implements the ApiKeyCreate interface.
 */
export function instanceOfApiKeyCreate(value: object): value is ApiKeyCreate {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function ApiKeyCreateFromJSON(json: any): ApiKeyCreate {
    return ApiKeyCreateFromJSONTyped(json, false);
}

export function ApiKeyCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiKeyCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'isActive': json['is_active'] == null ? undefined : json['is_active'],
        'expiresAt': json['expires_at'] == null ? undefined : (new Date(json['expires_at'])),
        'permissions': json['permissions'] == null ? undefined : json['permissions'],
    };
}

export function ApiKeyCreateToJSON(json: any): ApiKeyCreate {
    return ApiKeyCreateToJSONTyped(json, false);
}

export function ApiKeyCreateToJSONTyped(value?: ApiKeyCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'is_active': value['isActive'],
        'expires_at': value['expiresAt'] == null ? value['expiresAt'] : value['expiresAt'].toISOString(),
        'permissions': value['permissions'],
    };
}